distilled_model/requirements.txt
torch>=2.0
torchvision>=0.15
numpy
Pillow
matplotlib
scipy
tqdm
huggingface_hub
